using non-hd video, readme

+++ mikey/TokCog.py
@@ -9,7 +9,7 @@ headers = {
 	"X-RapidAPI-Key": RAPID_API_KEY,
 	"X-RapidAPI-Host": "tiktok-video-no-watermark2.p.rapidapi.com"
 }
-querystring = {"url":"https://www.tiktok.com/@tiktok/video/7231338487075638570","hd":"1"}
+querystring = {"url":"https://www.tiktok.com/@tiktok/video/7231338487075638570"}
 temp_filename = "tmp_tok.mp4"
 
 class TikTok(commands.Cog):
@@ -20,24 +20,28 @@ class TikTok(commands.Cog):
     async def on_message(self, message):
         if message.author.bot or message.content.startswith("!"):
             return
+        if os.path.exists(temp_filename):
+            os.remove(temp_filename)
         url = ""
-        print("tok triggered on " + message.content)
         for word in message.content.split(" "):
             if "tiktok.com" in word:
                 url = word
-                print("found tiktok - " + url)
         if url != "":
-            querystring["url"]=url
-            print("hitting api")
-            api_response = requests.get(api, headers=headers, params=querystring)
-            metadata = api_response.json()
-            video = requests.get(metadata["data"]["hdplay"], stream=True)
-            print("downloading video")
-            with open(temp_filename, "wb") as f:
-                for chunk in video.iter_content(chunk_size = 1024*1024):
-                    if chunk:
-                        f.write(chunk)
-            print("sending response")
-            await message.reply(file=File(temp_filename, filename="tiktok_file.mp4"))
-            print("removing file")
-            os.remove(temp_filename)
+            try:
+                print("started process")
+                querystring["url"]=url
+                api_response = requests.get(api, headers=headers, params=querystring)
+                metadata = api_response.json()
+                if metadata["code"] != 0:
+                    await message.reply(metadata["msg"])
+                    return
+                video = requests.get(metadata["data"]["play"], stream=True)
+                with open(temp_filename, "wb") as f:
+                    for chunk in video.iter_content(chunk_size = 1024*1024):
+                        if chunk:
+                            f.write(chunk)
+                await message.reply(file=File(temp_filename, filename="tiktok_file.mp4"))
+            except Exception as e:
+                await message.reply("unable to process video: " + e.message)
+            if os.path.exists(temp_filename):
+                os.remove(temp_filename)
